+++ libtiledbsoma/src/tiledbsoma/logger_public.h
