+++ libtiledbsoma/src/tiledbsoma/tiledbsoma
@@ -1,11 +1,11 @@
 /**
- * @file   stats.h
+ * @file   tiledbsoma
  *
  * @section LICENSE
  *
  * The MIT License
  *
- * @copyright Copyright (c) 2023 TileDB, Inc.
+ * @copyright Copyright (c) 2022 TileDB, Inc.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -27,23 +27,29 @@
  *
  * @section DESCRIPTION
  *
- *   This declares the common functions in the API
+ * This is the main import header for the C++ API
  */
 
-#ifndef TILEDBSOMA_STATS_H
-#define TILEDBSOMA_STATS_H
+#ifndef __TILEDBSOMA__
+#define __TILEDBSOMA__
 
-#include <stdexcept>  // for windows: error C2039: 'runtime_error': is not a member of 'std'
+// TODO Uncomment after finishing Python and R bindings
+// #include "soma_collection.h"
+// #include "soma_dataframe.h"
+// #include "soma_dense_ndarray.h"
+// #include "soma_experiment.h"
+// #include "soma_measurement.h"
+// #include "soma_object.h"
+// #include "soma_sparse_ndarray.h"
 
-#include <string>
+#include <cli/arrow_adapter.h>
+#include <tiledbsoma/logger_public.h>
+#include <utils/array_buffers.h>
+#include <utils/column_buffer.h>
+#include <utils/common.h>
+#include <utils/stats.h>
+#include <utils/version.h>
+#include <soma/managed_query.h>
+#include <soma/soma_array.h>
 
-namespace tiledbsoma::stats {
-
-__attribute__((visibility("default"))) void enable();
-__attribute__((visibility("default"))) void disable();
-__attribute__((visibility("default"))) void reset();
-__attribute__((visibility("default"))) std::string dump();
-
-};  // namespace tiledbsoma::stats
-
-#endif  // TILEDBSOMA_STATS_H
+#endif
