+++ libtiledbsoma/src/utils/column_buffer.h
@@ -38,8 +38,8 @@
 #include <span/span.hpp>
 #include <tiledb/tiledb>
 
-#include "tiledbsoma/common.h"
-#include "tiledbsoma/logger_public.h"
+#include "common.h"
+#include "logger.h"
 
 namespace tiledbsoma {
 
@@ -73,8 +73,7 @@ class ColumnBuffer {
      * @brief Convert a bytemap to a bitmap in place.
      *
      */
-    __attribute__((visibility("default"))) static void to_bitmap(
-        tcb::span<uint8_t> bytemap);
+    static void to_bitmap(tcb::span<uint8_t> bytemap);
 
     //===================================================================
     //= public non-static
