+++ libtiledbsoma/src/utils/logger.cc
